Add TF summary logging for validate loss
I know it's ugly but this serves the purpose I need. Having few data to work on, training loss and validate loss do not always go hand-in-hand, and this gives me an easy way to have an overview on how validation loss is performing (and infer how likely the model is to be successfully updated in the near future).