Change deprecated references to tf.compat.v1 counterparts
Since we're requiring tensorflow-1.14.0 I changed all the references that were giving out deprecation warnings. Those functions do the exact same thing as the "old" ones, so no actual code change is necessary.
This is more a graphic fix than a code fix, and they are not the only ones for sure, but it's a start.