Bypass multiprocessing.Pool.map for faster execution
After the introduction of multiprocessing management, inferring time has been really slow when working with a single image.
This workaround bypasses the pooling when inferring on a single image and using only one process.

There is for sure a way to implement this to cover more cases (single process, multiple images). But this is way less invasive and immediately effective.