Define custom Ludwig types and replace Dict[str, Any] type hints with them. (#2556)
This change is mostly motivated by readability as we use Dict[str, Any] type hints profusely but for a small set of very different objects. We are also soon introducing new libraries that will make ample use of FeatureDict and ModelConfigDict arguments.

This PR is less deep and broader than #2426 to cover other Dict[str, Any] objects that are passed around throughout the codebase like FeatureConfigDict, TrainingSetMetadataDict, and PreprocessingConfigDict.

When we define more formalized dataclasses for these other objects, like in #2426, it will be easy to identify all of the places that the richer dataclass types should be substituted.