Pass encoding dimensions to SequenceCombiner
If this argument isn't passed along then the `attention` reducer doesn't work for me. Note that although this fixes it for me I suspect there's lots of other places this would have to be applied. But maybe someone more experienced could chip in with if this is indeed an acceptable solution and other things to consider?